Add poll sensor Node-RED UI
Why:

- Allow sensors to be polled
- Allow displaying a sensors current value

This change addresses the need by:

- Update the Node-RED start sensor UI
- Fixing a run_until to duration_ms conversion bug
- Add reverse chronological ordering for data points
- Remove filtering by value of data points

+++ core/farms/models/controller_task.py
@@ -311,7 +311,7 @@ class ControllerTask(models.Model):
                 return {
                     "uuid": str(self.id),
                     "type": self.task_type,
-                    "duration_ms": (self.run_until - now).total_seconds() * 1000,
+                    "duration_ms": int((self.run_until - now).total_seconds() * 1000),
                     **self.parameters
                 }
             return {
